# Binary Reader performance improvement