*Description of changes:*
This PR is created to improve binary reader and parser performance. 

*Changes:* 

1. `IonBinaryReader`: removed duplicate maxId check from `getSymbolString()`.
2. `IonParserBinaryRaw`: Use bit shift for multiplication for `numberOfBytes` less than 4 in ` _readUnsignedIntAsNumberFrom()`.

*Test:*
Verified performance using [benchmarkjs](https://benchmarkjs.com) and [flame graphs](http://www.brendangregg.com/flamegraphs.html) from v8 profiling with large binary ion file as input.

*Test Results:*
Here times are measured using mean time of sample runs while Benchmarking(secs)
| Time(before enhancement)  | Possible enhancement | Time (after enhancement) | % decrease in Time
|----------|:-------------:|------:|-----:|
| 10.4261319358 | IonBinaryReader (removed duplicate condition checks) | 9.9510428 | 4.556715172274926
| 10.3857747234 | IonParserBinaryRaw (using bitshift for multiplication)| 10.3409678528 | 0.431425404395172
| 10.5577046098 | Both improvements together | 9.9062553234 | 6.170368564728586