@zslayton 2020-11-21
Per offline conversation, the `ion-test-driver` build is currently failing because `ion-test-driver` does an error message equality check on an `Error` that this PR changes.